normalize expr before applying dummy arg vars
it seems that they were appearing as "free" to `normalize`, and causing
it to crash without actually printing its error message.

since they are, in fact, "free" (we cheat and define them in the
interpreter `TermEnv` before running) it seems like the best thing to do
is to first normalize the `expr` body only, and then conjure the free vars
and apply them after.