tests work now
the main issue was that I foolishly copied the line

```
```

from the Holochain example tests
https://github.com/holochain/holochain/blob/develop/crates/holochain/tests/multi_conductor.rs

without realizing it made the test conditional on a flag that I didn't
have in my project. so the test never ran, which confused me. d'oh.

after that, had to resolve some version clashes. now the tests run, and
pass 👏